+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `error_from` macro copied from `anyhow::anyhow`
+- `bail` macro copied from `anyhow::bail`
+
 ## [0.1.2] - 2022-09-15
 
 ### Fixed